@@ -1,53 +1,40 @@
 {
-  "name": "bot-constructor",
-  "version": "4.0.0",
-  "scripts": {
-    "build": "react-scripts build",
-    "changelog": "auto-changelog --starting-version 3.0.0 --commit-limit false --hide-credit",
-    "eject": "react-scripts eject",
-    "lint": "eslint \"src/**/*.js\"",
-    "start": "react-scripts start",
-    "test": "react-scripts test",
-    "test:cov": "npm test -- --coverage --watchAll=false",
-    "test:debug": "react-scripts --inspect-brk test --runInBand"
-  },
-  "config": {
-    "coreui_library_short_version": "4.0"
-  },
+  "name": "temp-client-ui",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "@coreui/chartjs": "^3.0.0",
-    "@coreui/coreui": "^4.0.5",
-    "@coreui/icons": "^2.1.0",
-    "@coreui/icons-react": "2.0.0",
-    "@coreui/react": "4.0.0",
-    "@coreui/react-chartjs": "2.0.0",
-    "@coreui/utils": "^1.3.1",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.3",
-    "axios": "^0.24.0",
-    "antd": "^4.16.13",
-    "chart.js": "^3.5.1",
-    "classnames": "^2.3.1",
-    "core-js": "^3.18.3",
-    "enzyme": "^3.11.0",
-    "localforage": "^1.10.0",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-app-polyfill": "^2.0.0",
-    "react-dom": "^17.0.2",
-    "react-flow-renderer": "^9.6.11",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^5.3.0",
-    "redux": "4.1.1",
-    "simplebar-react": "^2.3.6"
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-scripts": "5.0.1",
+    "reactflow": "^11.11.4",
+    "web-vitals": "^2.1.4"
   },
-  "devDependencies": {
-    "auto-changelog": "~2.3.0",
-    "eslint": "^7.32.0",
-    "react-scripts": "^4.0.3",
-    "sass": "^1.43.2"
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  "engines": {
-    "node": ">=10",
-    "npm": ">=6"
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
