@@ -1,14 +1,8 @@
-import React from 'react'
-import { shallow } from 'enzyme/build'
-import App from './App'
-import Dashboard from './views/dashboard/Dashboard.js'
+import { render, screen } from '@testing-library/react';
+import App from './App';
 
-it('mounts App without crashing', () => {
-  const wrapper = shallow(<App />)
-  wrapper.unmount()
-})
-
-it('mounts Dashboard without crashing', () => {
-  const wrapper = shallow(<Dashboard />)
-  wrapper.unmount()
-})
+test('renders learn react link', () => {
+  render(<App />);
+  const linkElement = screen.getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
+});
